# Speed up cached ESLint checks