## What does this change?

Be more explicit about which folders each of the linting rule targets, and keep caches independent.

Only ignore the `static/js/curl-with-js-and-domReady.js` so `server/lib/get-content-from-url.js` and `webpack/webpack.config.dev-server.js` get linted!

## Why?

From 7-10s down to 1.5s on my machine this is very valuable for a task we want to run frequently for all.

## Screenshots

`hyperfine --warmup 2 "pnpm lint"`

| Before      | After      |
| ----------- | ---------- |
| 3.321 s ±  2.408 s | 2.056 s ±  0.044 s |

^ on my machine the result really are significantly better after… in a way that this benchmark does not make as clear