@mxdvl 2024-04-26
> Now that everything is linted I think we need to add the following playwright directories to `.eslintignore` (they are in .gitignore):
> 
> ```
> test-results/
> playwright-report/
> playwright/.cache/
> ```

PR welcome, I would have imagined these folders do not contain files that ESLint cares about?